pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policy-pilot"
version = "1.0.0"
description = "Offline AI-powered insurance policy document analyzer with semantic search and local LLM inference"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PolicyPilot Contributors", email = "dev@policypilot.local"}
]
keywords = [
    "nlp",
    "document-analysis",
    "semantic-search",
    "llm",
    "offline",
    "insurance",
    "policy",
    "faiss",
    "transformers"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic"
]

dependencies = [
    "faiss-cpu>=1.7.4",
    "flask>=3.0",
    "jsonschema>=4.0",
    "llama-cpp-python>=0.2.0",
    "numpy>=1.21",
    "PyPDF2>=3.0",
    "pytest>=7.0",
    "sentence-transformers>=2.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "isort>=5.12",
    "pytest-cov>=4.0",
    "pylint>=2.17",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/username/PolicyPilot"
Documentation = "https://github.com/username/PolicyPilot#readme"
Repository = "https://github.com/username/PolicyPilot.git"
"Bug Tracker" = "https://github.com/username/PolicyPilot/issues"

[tool.setuptools]
packages = [""]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
enable_error_code = "ignore-without-code"

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "*/__init__.py",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
